fix: landing build error
