+++ frontend/core/providers/helper.ts
@@ -20,7 +20,8 @@ import THEME from '~/const/theme'
 import METRIC from '~/const/metric'
 
 import { P } from '~/schemas'
-import { gqFetch, parseWallpaper, parseDashboard } from '~/utils/ssr'
+import { gqFetch } from '~/utils/api'
+import { parseWallpaper, parseDashboard } from '~/utils/ssr'
 import { extractQueryName } from '~/utils/graphql'
 import { loadLocaleFile } from '~/i18n'
 import { LOCALE } from '~/const/i18n'
