+++ frontend/landing/app/loading.tsx
@@ -1,9 +1,10 @@
 'use client'
 
-import LavaLampLoading from '~/widgets/Loading/LavaLampLoading'
+// import LavaLampLoading from '~/widgets/Loading/LavaLampLoading'
 
 const Loading = () => {
-  return <LavaLampLoading top={80} />
+  // return <LavaLampLoading top={80} />
+  return <div>loading ...</div>
 }
 
 export default Loading
